pyproject.toml
[project]
name = "strava-activity-cli"
version = "0.1.0"
description = "Command-line tool to manage Strava activities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "stravalib>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.scripts]
strava = "strava_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["strava_cli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
